requirements.txt
torch==2.1.0
torchaudio==2.1.0
openai-whisper>=20231117
pyautogui==0.9.54
pynput==1.7.6
sounddevice==0.4.6
numpy==1.24.3
pystray==0.19.4
Pillow==10.0.0  # Required for pystray
pyperclip>=1.8.2
scipy>=1.10.0
